combined camera and meal reducers
